feat(init): added the ability to intialize kakfka writers and close the connections gracefully when the server is closing

+++ internal/app/pkg/tokens/booking_token.go
@@ -102,7 +102,7 @@ func (bt *BookingToken) Get(
 	id = claims["sub"].(string)
 	driverID = claims["driver_id"].(string)
 	bookingID = claims["booking_id"].(string)
-	partitionNo = claims["partition_no"].(int)
+	partitionNo = int(claims["partition_no"].(float64))
 
 	return id, driverID, bookingID, partitionNo, nil
 }
